Some changes and bug fixes
Added experimental G5 (spline) support, fixed G93/G94 bug (sender issue #23) and updated G76 entry/exit taper handling - hopefully correct now (needs testing). Fixed some minor coding blunders.

+++ grbl/maslow.c
@@ -403,7 +403,7 @@ static bool maslow_segment_line (float *target, plan_line_data_t *pl_data, bool 
         do {
             idx--;
             delta[idx] = target[idx] - gc_state.position[idx];
-            max_delta = max(max_delta, fabs(delta[idx]));
+            max_delta = max(max_delta, fabsf(delta[idx]));
         } while(idx);
 
         if((segmented = !(pl_data->condition.rapid_motion || pl_data->condition.jog_motion) &&
